requirements.txt
streamlit==1.24.0
langchain==0.0.226
faiss-cpu==1.7.4
PyPDF2==3.0.1
openai==0.27.8
python-dotenv==1.0.0
